m3-9 fixup: migrate chat-only helpers from server.go to chat_proxy.go (CODE-8 file boundary cleanup)
